// src/app/generated/openapi/notip-management-api-openapi/model/update-tenant-request-dto.ts
/**
 * NoTIP Management API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface UpdateTenantRequestDto {
  name: string;
  status: UpdateTenantRequestDtoStatusEnum;
  suspension_interval_days: number;
}
export enum UpdateTenantRequestDtoStatusEnum {
  Active = 'active',
  Suspended = 'suspended',
}